requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-redis==4.10.0
djangorestframework==3.9.1
gunicorn==19.9.0
mysqlclient==1.4.2.post1
psycopg2==2.7.7
pytz==2018.9
redis==3.2.1
whitenoise==4.1.2
